完成对TestWindow的改造
